Use `Result` instead of `unwrap` in google code
`unwrap` is too eager to panic, so we get rid of all direct `unwrap`.
There is one `unwrap_or_else` block for some specific arithmetic that
is done, and that stays, but all regular unwraps which would panic
now return a result, which bubbles up to the proper channels.
On start, this panics, for a refresh, this is an error log.