feat: Use Google API Key over OAuth2 credentials
Makes using the app more useful, and now no longer requires a
webbrowser.

The API that couldn't use an API key was GoogleDocsV1.Document.Get,
this is replaced by a call to GoogleDriveV3.Files.Export, which exports
the document to plaintext. This means google can cache the exported
format, and the format does not need to be assembled as part of the API
calls. This requires minor changes to the parsing, but simplifies it
for the better.